Detach creation of experiment ID from start/stop
One might want to create an experiment (eg. configure sysData) for
an experiment to be executed, but not start the experiment straight
away. This allows a safety precaution by having a user manually
start the experiment via the UI, while still allowing remote
configuration of the experiment via the setSysData endpoint.